# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Bookstore API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/paket": {
            "get": {
                "tags": [
                    "Paket"
                ],
                "summary": "Mengambil semua data paket",
                "operationId": "89c95c6ed2282ba1dbe9680b2415bcba",
                "responses": {
                    "200": {
                        "description": "Daftar paket berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Paket"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Paket"
                ],
                "summary": "Menambahkan data paket baru",
                "operationId": "64fb1b04bffc98af60ea49ee407b679d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tracking_number",
                                    "weight",
                                    "status",
                                    "sender_id",
                                    "receiver_id",
                                    "item_name",
                                    "origin_address",
                                    "destination_address"
                                ],
                                "properties": {
                                    "tracking_number": {
                                        "type": "string"
                                    },
                                    "sender_id": {
                                        "type": "integer"
                                    },
                                    "receiver_id": {
                                        "type": "integer"
                                    },
                                    "item_name": {
                                        "type": "string"
                                    },
                                    "weight": {
                                        "type": "number"
                                    },
                                    "origin_address": {
                                        "type": "string"
                                    },
                                    "destination_address": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_transit",
                                            "delivered"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paket berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Paket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/paket/{id}": {
            "put": {
                "tags": [
                    "Paket"
                ],
                "summary": "Memperbarui status sebuah paket",
                "operationId": "507f9ee8ff8ff32ddc0a919ada63ae0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_transit",
                                            "delivered"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paket berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Paket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paket tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Paket"
                ],
                "summary": "Menghapus data paket",
                "operationId": "8cd434909a8a70214ed0d31b84b9522f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paket berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paket tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/paket/{tracking_number}/track": {
            "get": {
                "tags": [
                    "Paket"
                ],
                "summary": "Melacak paket berdasarkan nomor pelacakan",
                "operationId": "92f3da5e2112ad7c1f6b4794a96a4939",
                "parameters": [
                    {
                        "name": "tracking_number",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paket ditemukan dan data pelacakan berhasil ditampilkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Paket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paket tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Paket": {
                "required": [
                    "name",
                    "weight",
                    "destination"
                ],
                "properties": {
                    "id": {
                        "description": "ID Paket",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nama Paket",
                        "type": "string"
                    },
                    "weight": {
                        "description": "Berat Paket",
                        "type": "number",
                        "format": "float"
                    },
                    "destination": {
                        "description": "Tujuan Paket",
                        "type": "string"
                    },
                    "status": {
                        "description": "Status Pengiriman Paket",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}